# PERF: DataFrame(ndarray)