```
from asv_bench.benchmarks.frame_ctor import *
self = FromNDArray()
self.setup()

%timeit self.time_frame_from_ndarray()
37.9 µs ± 1.17 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- master
29.8 µs ± 549 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- PR
```